Refactor dynamic select to add builtin color palettes
