--- resource/get_language_list.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-from language_dict import SUPPORTED_LANGUAGES_SPACY
-
-
-def do(payload, config, plugin_config, inputs):
-    language_choices = sorted(
-        [{"value": k, "label": v} for k, v in SUPPORTED_LANGUAGES_SPACY.items()], key=lambda x: x.get("label")
-    )
-    language_choices.insert(0, {"label": "Language column", "value": "language_column"})
-    return {"choices": language_choices}
